SLIM compression version ready
Seems to be working now. Still doesn't handle rare potenial case of overflow in timestamp difference, but in back-end I check if the difference is 255, and if so I set the timestamp 1 less than reference timestamp. This is a neat solution, given that the overflow can never be more than -1 second.